moved arithmetic out of step.c
